rectangle shape for attachment
