+++ src/components/attachment-component.ts
@@ -1,4 +1,4 @@
-import { Attachment } from '../lib/attachment';
+import { Attachment } from '../lib/attachments/attachment';
 
 export class AttachmentComponent implements Phecs.Component {
   public attachments: Attachment[];
@@ -13,7 +13,7 @@ export class AttachmentComponent implements Phecs.Component {
     this.attachments = [];
   }
 
-  createAttachment(type: string, config: Attachment.Config): Attachment {
+  createAttachment(type: string, config: Attachments.Config): Attachment {
     const attachment = new Attachment(type, config, this.data.attachmentDebug, this.scene);
     this.attachments.push(attachment);
