+++ src/systems/has-hitboxes-system.ts
@@ -1,7 +1,7 @@
 import 'phaser';
 import { SpriteComponent } from '../components/sprite-component';
 import { AttachmentComponent } from '../components/attachment-component';
-import { Attachment } from "../lib/attachment";
+import { Attachment } from "../lib/attachments/attachment";
 import { HitboxComponent } from '../components/hitbox-component';
 import { EntityManager } from '../lib/phecs/entity-manager';
