+++ src/systems/has-hurtboxes-system.ts
@@ -2,7 +2,7 @@ import 'phaser';
 import { SpriteComponent } from '../components/sprite-component';
 import { HurtboxComponent } from '../components/hurtbox-component';
 import { EntityManager } from '../lib/phecs/entity-manager';
-import { Attachment } from '../lib/attachment';
+import { Attachment } from '../lib/attachments/attachment';
 import { AttachmentComponent } from '../components/attachment-component';
 
 export class HasHurtboxesSystem implements Phecs.System {
